Frontend: workaround for iOS splash screen logo
The logo on the ios version
of the splash screen has been too
large. Therefore, a new png that is
included only in the iOS project
has been added. This is a different file
than logo.png as it has more margin around
the cube. It is because of the maui bug that does
not let to resize logo on the splash screen.

Workaround for https://github.com/dotnet/maui/issues/11160